Clases
Se crearon las clases de cuadrado y circulo y se modifico el tipo de
dato del area de float a double para que aceptara los numeros con
decmimal

+++ Figuras/Figura.java
@@ -7,14 +7,14 @@
  */
 public class Figura
 {
-   protected float area;
+   protected double area;
    
    public Figura()
    {
        area = 0;
     }
     
-    public float accedeArea()
+    public double accedeArea()
     {
         return area;
     }
